+++ src/mame/msx/msx.h
@@ -163,6 +163,8 @@ protected:
 	u8 expanded_slot_r();
 	u8 kanji_r(offs_t offset);
 	void kanji_w(offs_t offset, u8 data);
+	u8 kanji2_r(offs_t offset);
+	void kanji2_w(offs_t offset, u8 data);
 	void ppi_port_a_w(u8 data);
 	void ppi_port_c_w(u8 data);
 	u8 ppi_port_b_r();
@@ -246,15 +248,15 @@ protected:
 	const XTAL m_main_xtal;
 	const int m_cpu_xtal_divider;
 
+	virtual void setup_slot_spaces(msx_internal_slot_interface &device);
+
 private:
 	// configuration helpers
 	template <typename T, typename U>
 	auto &add_base_slot(machine_config &config, T &&type, U &&tag, u8 prim, bool expanded, u8 sec, u8 page, u8 numpages, u32 clock = 0)
 	{
 		auto &device(std::forward<T>(type)(config, std::forward<U>(tag), clock));
-		device.set_memory_space(m_maincpu, AS_PROGRAM);
-		device.set_io_space(m_maincpu, AS_IO);
-		device.set_maincpu(m_maincpu);
+		setup_slot_spaces(device);
 		m_internal_slots.push_back({prim, expanded, sec, page, numpages, &device});
 		return device;
 	}
@@ -328,27 +330,15 @@ private:
 class msx2_base_state : public msx_state
 {
 protected:
-	msx2_base_state(const machine_config &mconfig, device_type type, const char *tag, XTAL main_xtal, int cpu_xtal_divider)
-		: msx_state(mconfig, type, tag, main_xtal, cpu_xtal_divider)
-		, m_v9938(*this, "v9938")
-		, m_v9958(*this, "v9958")
-		, m_rtc(*this, "rtc")
-		, m_rtc_latch(0)
-	{
-	}
+	msx2_base_state(const machine_config &mconfig, device_type type, const char *tag, XTAL main_xtal, int cpu_xtal_divider);
 
 	virtual void machine_start() override;
 
 	void msx2_base(ay8910_type ay8910_type, machine_config &config, const internal_layout &layout);
 	void msx2(ay8910_type ay8910_type, machine_config &config, const internal_layout &layout);
 	void msx2_pal(ay8910_type ay8910_type, machine_config &config, const internal_layout &layout);
-	void msx2plus_base(ay8910_type ay8910_type, machine_config &config, const internal_layout &layout);
-	void msx2plus(ay8910_type ay8910_type, machine_config &config, const internal_layout &layout);
-	void msx2plus_pal(ay8910_type ay8910_type, machine_config &config, const internal_layout &layout);
-	void turbor(ay8910_type ay8910_type, machine_config &config, const internal_layout &layout);
+	void msx2_v9958_base(ay8910_type ay8910_type, machine_config &config, const internal_layout &layout);
 	void msx2_add_softlists(machine_config &config);
-	void msx2plus_add_softlists(machine_config &config);
-	void turbor_add_softlists(machine_config &config);
 	void msx_ym2413(machine_config &config);
 	void msx2_64kb_vram(machine_config &config);
 	u8 rtc_reg_r();
@@ -356,19 +346,42 @@ protected:
 	void rtc_latch_w(u8 data);
 	u8 switched_r(offs_t offset);
 	void switched_w(offs_t offset, u8 data);
-	void turbo_w(int state);
 	void msx2_base_io_map(address_map &map);
 	void msx2_io_map(address_map &map);
-	void msx2plus_io_map(address_map &map);
+	void msx2_v9958_io_map(address_map &map);
 
 	std::vector<msx_switched_interface *> m_switched;
 
 	optional_device<v9938_device> m_v9938;
 	optional_device<v9958_device> m_v9958;
+	optional_device<ym2413_device> m_ym2413;
 	required_device<rp5c01_device> m_rtc;
 
 	// rtc
 	u8 m_rtc_latch = 0;
+};
+
+
+class msx2p_base_state : public msx2_base_state
+{
+protected:
+	msx2p_base_state(const machine_config &mconfig, device_type type, const char *tag, XTAL main_xtal, int cpu_xtal_divider);
+
+	void set_cold_boot_flags(u8 cold_boot_flags) { m_cold_boot_flags = cold_boot_flags; }
+
+	virtual void machine_start() override;
+
+	void msx2plus_base(ay8910_type ay8910_type, machine_config &config, const internal_layout &layout);
+	void msx2plus_pal_base(ay8910_type ay8910_type, machine_config &config, const internal_layout &layout);
+	void msx2plus(ay8910_type ay8910_type, machine_config &config, const internal_layout &layout);
+	void msx2plus_pal(ay8910_type ay8910_type, machine_config &config, const internal_layout &layout);
+	void msx2plus_io_map(address_map &map);
+	void msx2plus_add_softlists(machine_config &config);
+	void turbor_add_softlists(machine_config &config);
+	void turbor(ay8910_type ay8910_type, machine_config &config, const internal_layout &layout);
+
+	u8 m_cold_boot_flags;
+	u8 m_boot_flags;
 };
 
 #endif // MAME_MSX_MSX_H
